consolidate methods to base
