+++ warn_classes/fl.py
@@ -65,27 +65,3 @@ class FLWarn(Warn):
             return dfs
         except:
             return None
-        
-    def get_month_date_year(self, date:str):
-        date_regex = re.compile(r'(\d{1,2})/(\d{1,2})/(\d{1,4})')
-        match = date_regex.search(date)
-        if match:
-            month = match.group(1)
-            date = match.group(2)
-            year = match.group(3)
-
-            if len(month) == 1:
-                month = "0" + month
-            if len(date) == 1:
-                date = "0" + date
-
-        return month, date, year
-    
-    def get_company_name(self, text):
-        match = re.match(r"^\d+\s*(.*)$", text)
-
-        if match:
-            captured_text = match.group(1)
-            return captured_text 
-        else:
-            return None
